refactor: extract the config model out of the RouteManager god class (US-013)
The configuration data model (Config, RoutingMode, VPNType, ProxyConfig, DomainEntry,
ServiceEntry) was nested inside the 4.5k-line @MainActor RouteManager, so the "pure"
CommandRouter/CLI and the rule/route builders had to reach into the god class's
namespace (RouteManager.Config appears 30x in CommandRouter alone) — a layering
inversion, and a blocker to ever making the model its own leaf module.

Move those types to top-level in ConfigModel.swift (pure Codable value types, no
RouteManager, no side effects). RouteManager keeps thin `typealias Config = ...`
aliases so the historical `RouteManager.X` spelling across the codebase keeps working
with zero reference churn.

Behavior-preserving: identical Codable shape (JSON keys + defaults unchanged), so
existing configs load exactly as before. Build clean, 786 tests / 0 failures.
RouteManager.swift 4455 -> 3956 lines.
